implemented orb_functions.py in /analyze endpoint
