+++ docker/orb.requirements.txt
@@ -13,3 +13,4 @@ ruptures
 mpld3
 python-multipart
 influxdb
+prophet
